+++ packages/jira-adapter/src/change_validators/index.ts
@@ -34,7 +34,7 @@ import { automationsValidator } from './automations'
 import { lockedFieldsValidator } from './locked_fields'
 import { globalProjectContextsValidator } from './global_project_contexts'
 import { systemFieldsValidator } from './system_fields'
-import { workflowPropertiesValidator } from './workflow_properties'
+import { workflowPropertiesValidator } from './workflows/workflow_properties'
 import { permissionSchemeValidator } from './sd_portals_permission_scheme'
 import { wrongUserPermissionSchemeValidator } from './wrong_user_permission_scheme'
 import { GetIdMapFunc } from '../users_map'
