+++ packages/jira-adapter/src/change_validators/workflows/workflow_properties.ts
@@ -16,9 +16,9 @@
 import _ from 'lodash'
 import { ChangeValidator, getChangeData, isInstanceChange, SeverityLevel } from '@salto-io/adapter-api'
 import { collections } from '@salto-io/lowerdash'
-import { isWorkflowInstance, Status, Transition } from '../filters/workflow/types'
-import { Values } from '../../../adapter-api/src/values'
-import { WORKFLOW_TYPE_NAME } from '../constants'
+import { isWorkflowInstance, Status, Transition } from '../../filters/workflow/types'
+import { Values } from '@salto-io/adapter-api/src/values'
+import { WORKFLOW_TYPE_NAME } from '../../constants'
 
 
 const { awu } = collections.asynciterable
